requirements.txt
flask

flask-SocketIO

adafruit_motorkit
sense_hat
adafruit_ads1x15.analog_in